pgwire: update generate-binary tool; use in new tests
cmp-protocol is a new tool to compare postgres and cockroach pgwire
encodings. Statements can be specified on the command line or generated
randomly. pgx is used to stuff bytes over a connection manually.

A new pgconnect package abstracts away the connection
handling. generate-binary (which was previously broken on new postgres
servers) has been updated to use this new package.

In addition, its output has been changed to output to a single file
instead of many. This allows the previous encoding tests to be removed
in favor of a single test. We simplify it by using our parser to convert
a SQL expression to a datum instead of hard coding a type -> parser
func as before, which should allow for more types to be added in the
future. These new tests cover exactly the same cases as the previous ones.

Release note: None

wip